% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atest.R
\name{atest}
\alias{atest}
\title{add Test-statement to predefined dataframe}
\usage{
atest(object, expect, description, test_df = test_dataframe, assign = TRUE)
}
\arguments{
\item{object}{The object to be testet (should be a vector). In case of a list only first element will be used for evaluation.}

\item{expect}{Character. Statement to be evaluated}

\item{description}{A description of the test. Must be unique for all collected tests within the dataframe.}

\item{test_df}{Specify dataframe where evaluation should be added.}

\item{assign}{Logical. Choose TRUE to add the evaluation to the test_df.}
}
\description{
add Test-statement to predefined dataframe
}
\examples{
atest(iris$Species, "\%in\% c('setosa', 'virginica', 'versicolor')")
}
\seealso{
\code{\link[fsf]{utest}}, \code{\link[fsf]{stest}}
}
