% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/audioplayer_set_path.R
\name{audioplayer_set_path}
\alias{audioplayer_set_path}
\title{wrapper for tuneR::setWavPlayer}
\usage{
audioplayer_set_path(path2player = "/usr/bin/play")
}
\arguments{
\item{path2player}{Character. Specify path to audioplayer}
}
\description{
wrapper for tuneR::setWavPlayer
}
