% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backupfile.R
\name{backupfile}
\alias{backupfile}
\title{backup a file}
\usage{
backupfile(filename = "", filepath = NULL)
}
\arguments{
\item{filename}{Character. Specify description for filepath. Will be appended after original filename}

\item{filepath}{Character. Specify file path you want to copy.}
}
\description{
backup a file
}
