% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birdlife_remo.R
\name{birdlife_remo}
\alias{birdlife_remo}
\title{play random recording from Remo Probst from list of birds (birdlife course)}
\usage{
birdlife_remo()
}
\description{
play random recording from Remo Probst from list of birds (birdlife course)
}
