% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/c2txt.R
\name{c2txt}
\alias{c2txt}
\title{Convert vector as text}
\usage{
c2txt(vector, frame = "\\"", separator = ", ")
}
\arguments{
\item{vector}{Vector that should be represented as text.}

\item{frame}{Character. Specify character that should be printed before and after each element of the vector.}

\item{separator}{Character. Specify character that should be printed between each element of the vector.}
}
\value{
Text representation of the vector.
}
\description{
Convert vector as text
}
\examples{
c2txt(names(iris)); c2txt(paste0(letters, "='", 1:length(letters), "'"), separator = ",", frame = "")
}
\author{
Frederik Sachser
}
