% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmt.R
\name{cmt}
\alias{cmt}
\title{insert docx comment}
\usage{
cmt(highlight = "", comment, name = Sys.info()["user"], thedate = Sys.time())
}
\arguments{
\item{highlight}{Character. Text that should be higlighted (to comment on).}

\item{comment}{Character. The Comment inside the box.}

\item{name}{Character. Name of coment author.}

\item{thedate}{Character. Date of the comment.}
}
\description{
insert docx comment
}
