% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comcomcol.R
\name{comcomcol}
\alias{comcomcol}
\title{COMpare COMmon COLumns of multiple dataframes}
\usage{
comcomcol(..., custom_fun = class, freq_only = FALSE,
  freq_sort = TRUE)
}
\arguments{
\item{...}{Objects of class data.frame.}

\item{custom_fun}{object of class function. `comcomcol` acts as a functional and takes another function as input (wrapper for lapply). Specify any function for comparison. Default is `class`.}

\item{freq_only}{Logical. Specify if custom_fun should be included.}

\item{freq_sort}{Logical. Specify if output should be sorted according to frequency of common columns.}
}
\description{
This function will count and print the number of common columns of given dataframes and check consistency of additional values. Comparison is based on colnames. For single dataframes the output is per default a representation of `lapply(<dataframe>, class)` (formatted as a dataframe).
}
\author{
Frederik Sachser
}
