% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curfp.R
\name{curfp}
\alias{curfp}
\title{Get current file path}
\usage{
curfp()
}
\description{
This function will determine and return the path of the executing script.
}
\note{
'curfp' stands for 'current file path'.
}
\seealso{
\url{http://stackoverflow.com/a/36777602/7613841}, \url{http://stackoverflow.com/a/32016824/2292993}, \url{http://stackoverflow.com/a/35842176/2292993}, \url{http://stackoverflow.com/a/15373917/7613841}, \url{http://stackoverflow.com/a/1815743/7613841}, \url{http://stackoverflow.com/a/1816487/7613841}
}
\author{
Jerry T, aprstar, Richie Cotton, steamer25, Suppressingfire, hadley, Frederik Sachser
}
