% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2list.R
\name{data2list}
\alias{data2list}
\title{Combine multiple dataframes into a named list}
\usage{
data2list(df)
}
\arguments{
\item{df}{character. Character vector of dataframe-objects.}
}
\description{
Combine multiple dataframes into a named list
}
\examples{
data2list(c("iris", "mtcars"))
}
