% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman_as_dataframe.R
\name{dataman_as_dataframe}
\alias{dataman_as_dataframe}
\title{create a template for a data manual for multiple dataframes}
\usage{
dataman_as_dataframe(named_list_of_dataframes)
}
\arguments{
\item{named_list_of_dataframes}{object of class 'list' with named elements containing dataframes}
}
\description{
rows merged by variable names
}
\examples{
dataman_as_dataframe(data2list(c("chickwts", "ChickWeight")))
}
