% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman_as_dataframe_skim.R
\name{dataman_as_dataframe_skim}
\alias{dataman_as_dataframe_skim}
\title{create a template for a data manual for multiple dataframes}
\usage{
dataman_as_dataframe_skim(named_list_of_dataframes)
}
\arguments{
\item{named_list_of_dataframes}{}
}
\description{
rows separated by dataframe
}
\examples{
dataman_as_dataframe(data2list(c("chickwts", "ChickWeight")))
}
