% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataman_template.R
\name{dataman_template}
\alias{dataman_template}
\title{print template for data manual}
\usage{
dataman_template(data)
}
\arguments{
\item{data}{Object of class data.frame.}
}
\description{
print roxygen2 template for data documentation.
}
