% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dircomp.R
\name{dircomp}
\alias{dircomp}
\title{compare files in directories}
\usage{
dircomp(dir1, dir2)
}
\arguments{
\item{dir1}{Character. Specify path to directory 1.}

\item{dir2}{Character. Specify path to directory 2.}
}
\description{
compare files in directories
}
