% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dstrip.R
\name{dstrip}
\alias{dstrip}
\title{density strip with some defaults}
\usage{
dstrip(jagsout, parameters_index)
}
\arguments{
\item{jagsout}{object oc class jagsUI.}

\item{parameters_index}{integer vector. Specify index for parameters (intended for output of output from jagsUI).}
}
\description{
density strip with some defaults
}
\seealso{
\code{\link[denstrip]{denstrip}}
}
