% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fe.R
\name{fe}
\alias{fe}
\title{write custom Rmd-template or open existing file}
\usage{
fe(file, ...)
}
\arguments{
\item{file}{Character. Specify full file path to open or create a file.}

\item{...}{Arguments passed to 'utils::file.edit'.}
}
\description{
This function will open existing files or create a new one.
}
\seealso{
\code{\link[utils]{file.edit}}
}
\author{
Frederik Sachser
}
