% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_nonASCII.R
\name{find_nonASCII}
\alias{find_nonASCII}
\title{find Non-ASCII characters within a file}
\usage{
find_nonASCII(filepath)
}
\arguments{
\item{filepath}{Character. Specify path to the file that you want to check for Non-ASCII characters.}
}
\description{
find Non-ASCII characters within a file
}
\note{
use sapply to search for non ASCII characters in multiple files
}
