% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findlines.R
\name{findlines}
\alias{findlines}
\title{Find lines in a bunch of files}
\usage{
findlines(
  expression = "# magic word",
  exclusive = FALSE,
  path = ".",
  filepattern = "\\\\.R$|\\\\.Rmd$",
  recursive = TRUE
)
}
\arguments{
\item{expression}{Character. Specify expression to match the line.}

\item{exclusive}{Logical. If TRUE use which, else grep.}

\item{path}{Character. Specify path to the directory including the files.}

\item{filepattern}{Character. Specify file pattern.}

\item{recursive}{Logical.}
}
\description{
Works like grep or which for a bunch of files (line by line). Useful example: 1. Always use the same predefined word(s) to mark positions in scripts where you should review something at some time (e.g. 'TODO'). 2. Search for those words within a bunch of files to get an overview.
}
\note{
If exclusive == FALSE you can use regular expressions (and therefore multiple expressions at a time).
}
\author{
Frederik Sachser
}
