% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggtraceplot.R
\name{ggtraceplot}
\alias{ggtraceplot}
\title{Traceplot for jagsUI objects}
\usage{
ggtraceplot(jagsUI_object, filepath = NULL, title)
}
\arguments{
\item{jagsUI_object}{Object of class jagsUI.}

\item{filepath}{Optional. Specify filepath (including adequate extension) in case you want to save the figure. Printing will be supressed.}

\item{mytitle}{Optional. Specify title.}
}
\description{
Traceplot for jagsUI objects
}
