% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_by_diff.R
\name{group_by_diff}
\alias{group_by_diff}
\title{create groups by numerical difference threshold}
\usage{
group_by_diff(x, difference = 1)
}
\arguments{
\item{x}{numeric. Vector that should be binned into groups. Also possible with Dates.}

\item{difference}{Integer. Specify maximum within-group difference.}
}
\value{

}
\description{
create groups by numerical difference threshold
}
\note{
vector should be arranged (sorted) prior to group_by_diff.
}
