% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hashtag.R
\name{hashtag}
\alias{hashtag}
\title{insert hashtag}
\usage{
hashtag()
}
\description{
function for RStudio addin. Assign it to some custom shortcut to insert 'a hashtag comment' to your current cursor position.
}
