% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/headr.R
\name{headr}
\alias{headr}
\title{insert preformatted header in Rscript-style}
\usage{
headr(mytext = "this is random text")
}
\arguments{
\item{mytext}{Character. Specify text for the header.}
}
\description{
insert preformatted header in Rscript-style
}
