% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ics2table.R
\name{ics2table}
\alias{ics2table}
\title{Convert ics files to data.frame}
\usage{
ics2table(calendar_ics, calname = NULL)
}
\arguments{
\item{calendar_ics}{Character. Path (or URL) to ics-file.}

\item{calname}{Character. If specified an additional column will be written containing the calname (might be useful for further processing).}
}
\value{
data.frame
}
\description{
Convert ics files to data.frame
}
\note{
Reoccurring events are not supported!
}
\author{
Frederik Sachser
}
