% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img2pdfmerge.R
\name{img2pdfmerge}
\alias{img2pdfmerge}
\title{merge multiple image files into one pdf}
\usage{
img2pdfmerge(thedir = ".")
}
\arguments{
\item{thedir}{Character. Specify file path of the directory containing image files (e.g. jpg, gif,...)}
}
\description{
files should be named 1.gif, 2.gif and so on. output will be named according to the directory name
}
