% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_cmt.R
\name{incmt}
\alias{incmt}
\title{insert docx comment}
\usage{
incmt()
}
\description{
function for RStudio addin. Assign it to some custom shortcut to insert a docx comment to yout current cursor position.
}
