% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_mark.R
\name{insert_mark}
\alias{insert_mark}
\title{insert html mark}
\usage{
insert_mark()
}
\description{
function for RStudio addin. Assign it to some custom shortcut to insert <mark></mark> to yout current cursor position.
}
