% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jagsprint.R
\name{jagsprint}
\alias{jagsprint}
\title{trim summary output from jagsUI}
\usage{
jagsprint(
  jagsobject,
  summaryrows = 1:20,
  digi = 2,
  summarycolumns = c(1, 3, 7:10)
)
}
\arguments{
\item{jagsobject}{Specify object of class jagsUI you want to print.}

\item{summaryrows}{Integer vector. Specify rows of summary table of the object.}

\item{digi}{Integer. Specify digits to trim the output.}

\item{summarycolumns}{Integer vector. Specify columns of summary table of the object.}
}
\description{
trim summary output from jagsUI
}
