% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{mess}
\alias{mess}
\title{clean the console, leave a message}
\usage{
mess(message = "\\v\\017\\v\\v\\v\\017\\v\\v\\017 \\nJazz up your day")
}
\arguments{
\item{message}{Character Specify a message}
}
\description{
clean the console, leave a message
}
