% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nasum.R
\name{nasum}
\alias{nasum}
\title{NA sum}
\usage{
nasum(dataframe)
}
\arguments{
\item{dataframe}{Object of class dataframe.}
}
\description{
obtain the rows and columns of a dataframe, that contain missing data only
}
