% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{note}
\alias{note}
\title{Create a template Rmd-file for taking notes}
\usage{
note(filename = "note.Rmd", dir = "default", add_date = TRUE,
  Author = Sys.info()["effective_user"], Date = "`r Sys.Date()`")
}
\arguments{
\item{filename}{Character. Specify the filename to save the new file.}

\item{dir}{Character. Specify directory where to save the new file. Default
will be "Schreibtisch" or "Desktop".}

\item{add_date}{Logical. If TRUE the filename will be complemented by the
current date.}

\item{Author}{Character. Optionally customize the name of the Author (used
for the YAML header). Default is the effective user of the system info.}

\item{Date}{Character. Optionally customize the date (used for the YAML
header). Default is the current Date (format YYYY-MM-DD).}
}
\description{
This function will create an Rmd-file including some predefined
  lines. The function is designed to save some time writing the same input
  again and again (e.g. Author, date and other things).
}
\note{
Missing directories will be created recursively.

It is not possible to overwrite existing files.

Other YAML header options will be choosen automatically. Edit the
  resulting file to customize the YAML header.
}
\author{
Frederik Sachser
}

