% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairex.R
\name{pairex}
\alias{pairex}
\title{graphics::pairs example}
\usage{
pairex(dataframe, ...)
}
\arguments{
\item{dataframe}{Object of class dataframe. Subset will be applied automatically (numeric only).}

\item{...}{pass arguments to panel.hist and panel.cor}
}
\description{
graphics::pairs example
}
