% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par1.R
\name{par1}
\alias{par1}
\title{par temporary}
\usage{
par1(theplots, ...)
}
\arguments{
\item{...}{specify parameters of par().}

\item{theplots.}{List of plots that should be processesed}
}
\description{
par temporary
}
\examples{
par1(plot(glm(Sepal.Width ~ ., data = iris)), mfrow = c(2,2))
}
