% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasterisk.R
\name{pasterisk}
\alias{pasterisk}
\title{Classify significance levels
Print asterisk}
\usage{
pasterisk(p.value)
}
\arguments{
\item{p.value}{Numeric vector. Specify vector to classify.}
}
\description{
Classify significance levels
Print asterisk
}
\author{
Frederik Sachser
}
