% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pc.R
\name{pc}
\alias{pc}
\title{Paste Character}
\usage{
pc(char = "-", times = 80)
}
\arguments{
\item{char}{Character. Specify Character to paste.}

\item{times}{Repeat character n times. Default = 80}
}
\description{
Paste Character
}
