% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_copy2root.R
\name{pdf_copy2root}
\alias{pdf_copy2root}
\title{Copy pdf files from subdirectories to root dir}
\usage{
pdf_copy2root(path)
}
\arguments{
\item{path}{Character. Specify root directory.}
}
\description{
Copy pdf files from subdirectories to root dir
}
