% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdftk_merge.R
\name{pdftk_merge}
\alias{pdftk_merge}
\title{merge pdfs within dir}
\usage{
pdftk_merge(path)
}
\arguments{
\item{path}{Character. File path to directory containing pdfs.}
}
\description{
merge pdfs within dir
}
