% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdata_ls.R
\name{pkgdata_ls}
\alias{pkgdata_ls}
\title{Return vector of data within package}
\usage{
pkgdata_ls(pkg)
}
\arguments{
\item{pkg}{Character. Specify Name of an installed package.}
}
\description{
Contrary to utils::data the function will only return a character vector of data within a package.
}
\seealso{
\code{\link[fsf]{pkgdata_yank}}, \code{\link[fsf]{pkgdata_write}}
}
\author{
Frederik Sachser
}
