% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdata_write.R
\name{pkgdata_write}
\alias{pkgdata_write}
\title{Write dataframe}
\usage{
pkgdata_write(listofdf, target_dir = "inst/extdata")
}
\arguments{
\item{listofdf}{Character. Character vector of the data frame object(s).}

\item{target_dir}{Character. Specify target directory (default 'inst/extdata')}
}
\description{
This function provides a fast way to write multiple dataframes into a directory. File extension will be RData.
}
\note{
directory will be written recursively and existing files of the same name will be overwritten.
}
\seealso{
\code{\link[sf]{pkgdata_ls}}, \code{\link[sf]{pkgdata_yank}}
}
