% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgdata_yank.R
\name{pkgdata_yank}
\alias{pkgdata_yank}
\title{Copy all dataframes of an installed R-package}
\usage{
pkgdata_yank(pkg, target_dir = "inst/extdata", listofdf = NULL)
}
\arguments{
\item{pkg}{Character. Name of the installed package.}

\item{target_dir}{Character. Specify target directory for dataframes (one RData-file for each dataframe).}

\item{listofdf}{Character. Specify subset of dataframes.}
}
\description{
Copy all dataframes of an installed R-package
}
\seealso{
\code{\link[fsf]{pkgdata_ls}}, \code{\link[fsf]{pkgdata_write}}
}
