% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDataload.R
\name{RDataload}
\alias{RDataload}
\title{load RData-files to globalenv}
\usage{
RDataload(path = "inst/extdata")
}
\arguments{
\item{path}{Character. Specify file path where RData-files are stored.}
}
\description{
load all RData-files within a directory (recursive) into globalenv.
}
\author{
Frederik Sachser
}
