% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readc.R
\name{readc}
\alias{readc}
\title{import many files using rio::import}
\usage{
readc(files)
}
\arguments{
\item{files}{character vector. Specify multiple file paths of data for import.}
}
\description{
import many files using rio::import
}
\seealso{
\code{\link[rio]{import}}
}
\author{
Frederik Sachser
}
