% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_std.R
\name{rename_std}
\alias{rename_std}
\title{Rename df columns}
\usage{
rename_std(.data)
}
\arguments{
\item{.data}{An object of class data.frame}
}
\description{
rename all variables within a data frame (standardized syntax).
}
\references{
see https://www.r-bloggers.com/clean-consistent-column-names/
}
