% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowid_before_after.R
\name{rowid_before_after}
\alias{rowid_before_after}
\title{Check if row id did not change after some manipulation task}
\usage{
rowid_before_after(a, b, rowid = "rowid")
}
\arguments{
\item{a}{Dataframe or vector of row id (after manipulation).}

\item{b}{Dataframe or vector of row id (before manipulation).}

\item{rowid}{Character. Name of variable if a/b is a dataframe.}
}
\description{
Check if row id did not change after some manipulation task
}
\author{
Frederik Sachser
}
