% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdf.R
\name{sdf}
\alias{sdf}
\title{Structure as a data frame}
\usage{
sdf(data, n = 3)
}
\arguments{
\item{data}{Object of class data.frame.}

\item{n}{Number of rows to be printed.}
}
\description{
similar to utils::str(), but organized in a dataframe. Additionally to the class the amount of NA-values is returned.
}
\author{
Frederik Sachser
}
