% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skimview.R
\name{skimview}
\alias{skimview}
\title{skim data in excel}
\usage{
skimview(dataframe, xview = TRUE)
}
\arguments{
\item{dataframe}{Object of class dataframe}

\item{xview}{logical. TRUE will open the file via xview.}
}
\description{
Wrapper for skimr::skim and fsf::xview
}
\seealso{
\code{\link[fsf]{xview}}
}
