% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourcermd.R
\name{sourcermd}
\alias{sourcermd}
\title{convert Rmd to R and source}
\usage{
sourcermd(rmd_file = NULL)
}
\arguments{
\item{rmd_file}{Character. Specify file_path to Rmd-document}
}
\description{
convert Rmd to R and source
}
