% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stest.R
\name{stest}
\alias{stest}
\title{print Summary of test_df}
\usage{
stest(test_df = test_dataframe, return = TRUE)
}
\arguments{
\item{test_df}{A dataframe of specific format (output from sf::atest).}

\item{return}{Logical. Specify if test_df should be returned.}
}
\description{
print Summary of test_df
}
\seealso{
\code{\link[sf]{atest}}
}
