% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplitunnest.R
\name{strsplitunnest}
\alias{strsplitunnest}
\title{Split a string into multiple rows}
\usage{
strsplitunnest(dataframe, columnname, separator = ",", affix = "_split")
}
\arguments{
\item{dataframe}{Object of class "data.frame" or tibble.}

\item{columnname}{Character. Name of the column containing a string for splitting.}

\item{separator}{Character. Specify the separator that should be used to split the string.}

\item{affix}{Character. Specify affix for the new column.}
}
\description{
Split a string into multiple rows
}
\author{
Frederik Sachser
}
