% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template4articles.R
\name{template4articles}
\alias{template4articles}
\title{create Rmd with predefined text for quick notes on articles}
\usage{
template4articles(file = file.path(tempdir(), "example.Rmd"), ...)
}
\arguments{
\item{file}{Character. Specify full file path to open or create a file.}

\item{...}{Arguments passed to 'utils::file.edit'.}
}
\description{
This function will open existing files or create a new one with predefined text.
}
\seealso{
\code{\link[utils]{file.edit}}, \code{\link[fsf]{fe}}
}
\author{
Frederik Sachser
}
