% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tprint.R
\name{tprint}
\alias{tprint}
\title{trim print output}
\usage{
tprint(object2print, maxprint = 100, digi = 2)
}
\arguments{
\item{object2print}{Object that should be printed.}

\item{maxprint}{Integer. Specify for temporary change of max.print option. Choose 0 (zero) to set maximum possible.}

\item{digi}{Integer. Specify digits to trim the output.}
}
\description{
trim print output
}
