% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/txt_pre_ext.R
\name{txt_pre_ext}
\alias{txt_pre_ext}
\title{insert text before extension of filename}
\usage{
txt_pre_ext(filename, txt_pre_ext = now())
}
\arguments{
\item{filename}{Character. Specify filename (including extension).}

\item{txt_pre_ext}{Character. Specify text to paste right before the file extension.}
}
\description{
insert text before extension of filename
}
\author{
Frederik Sachser
}
