% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utest.R
\name{utest}
\alias{utest}
\title{Undo last Test}
\usage{
utest(test_df = test_dataframe)
}
\arguments{
\item{test_df}{A dataframe of specific format (output from fsf::atest).}
}
\description{
Undo last Test
}
\seealso{
\code{\link[fsf]{atest}}, \code{\link[fsf]{stest}}
}
