% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_data.R
\name{write_data}
\alias{write_data}
\title{Write a bunch of dataframes}
\usage{
write_data(
  listofdf = "GlobalEnv",
  target_dir_data = "data",
  data_extension = "RData",
  rebuild_target_dir_data = FALSE,
  summarize_df = TRUE
)
}
\arguments{
\item{listofdf}{Character vector. Specify a vector of dataframes to write. Choose 'GlobalEnv' to write all data frames currently inside your global environment.}

\item{target_dir_data}{Character. Target directory for data.}

\item{data_extension}{Character. Specify file extension for data frames. Possible values: 'csv', 'RData' and 'rds'.}

\item{rebuild_target_dir_data}{Logical. Delete target_dir_data before writing current data frames.}

\item{summarize_df}{Logical. Should a list of written data frames be printed to console?}
}
\description{
This function will write multiple dataframes into separate files.
}
\author{
Frederik Sachser
}
