% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataframes_to_csv.R
\name{write_dataframes_to_csv}
\alias{write_dataframes_to_csv}
\title{Write multiple dataframes to a single CSV file}
\usage{
write_dataframes_to_csv(df_list, file)
}
\arguments{
\item{df_list}{list of dataframes}

\item{file}{filename}
}
\description{
Write multiple dataframes to a single CSV file
}
\details{
This function was originally developed by an unkown author (github-username 'd-notebook') and hosted via github. The repository 'sheetr' and associated account 'd-notebook' has been removed recently. Because another function (xview) depends on the code and the package is not available any longer I decided to include it directly within this package.
}
\author{
d-notebook
}
