% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xc.R
\name{xc}
\alias{xc}
\title{Preformatted text to embed xeno-canto recordings}
\usage{
xc(kat_nr)
}
\arguments{
\item{kat_nr.}{Integer. Specify Kat. Nr. of xeno-canto recording.}
}
\description{
Preformatted text to embed xeno-canto recordings
}
