% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xflex.R
\name{xflex}
\alias{xflex}
\title{Write and View dataframe as flextable html}
\usage{
xflex(table, target = "/tmp/foo.html", open = TRUE)
}
\arguments{
\item{table}{Object of class data.frame.}

\item{target}{Specify target filepath for the html-table.}

\item{open}{logical. Open document if open = TRUE (default).}
}
\description{
Write and View dataframe as flextable html
}
\seealso{
\code{\link[flextable]{save_as_html}}
}
