% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xggsave.R
\name{xggsave}
\alias{xggsave}
\title{Externally open tiff-file after writing via ggsave}
\usage{
xggsave(filename = NULL, ...)
}
\arguments{
\item{...}{specify further arguments for ggsave}

\item{copy2target}{Character. Optionally specify file path to copy the resulting file}
}
\description{
Externally open tiff-file after writing via ggsave
}
