% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xview.R
\name{xview}
\alias{xview}
\title{External View}
\usage{
xview(
  x,
  file = paste0("xview_", now(), ".xlsx"),
  deldep = TRUE,
  list_collapse = TRUE,
  open_doc = TRUE,
  ...
)
}
\arguments{
\item{x}{Object to export (e.g. data.frame)}

\item{file}{Specify filename. Directories will not be created.}

\item{deldep}{Logical. Specify if deprecated files should be deleted. works only if file_path begins with "xview_" followed by 14 integer values (default for exporting new files).}

\item{list_collapse}{Logical. Collapse list of dataframes? (depends on github-package d-notebook/sheetr)}

\item{open_doc}{Logical. Should document be opened afterwards?}

\item{...}{Further params passed to rio::export (e.g. overwrite, row.names, etc.)}
}
\description{
This function will export an object to a file (using package "rio"). A system command is invoked to open the file afterwards.
}
\seealso{
\code{\link[rio]{export}}, \code{\link[sheetr]{write_dataframes_to_csv}}
}
\author{
Frederik Sachser
}
