% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ztrans.R
\name{ztrans}
\alias{ztrans}
\title{simple Z transformation}
\usage{
ztrans(x)
}
\arguments{
\item{x}{Numeric. Input vector for z transformation}
}
\description{
simple Z transformation
}
\author{
Frederik Sachser
}
